Added the structure for a basic bot.
Turns out that a connection to a server is the only granularity of control you
get when connecting to IRC. Servers are likely some sort of pub-sub model. So,
in light of this, I have distributed functionality from IRC/Server to
IRC/BasicBot and IRC/Conn and removed the IRC/Server module.