Changed test.pl to be zasz_driver.
I have decided to name my bot Zasz, a worthy cerebrate of the hive.

BasicBot is a basic bot implementation that is able to connect (using
IRC::Conn) to an IRC server and channels, and PONG to PINGs.

ZaszBot is going to be my NLP implementation of my bot, and likely extended to
be my IRC Bot of the future.